separate validator cards basic / detailed
